fix(agent): preserve reasoning order after hydration
