+++ packages/shared/src/in-app-agent/index.ts
@@ -5,3 +5,4 @@ export * from "./schema";
 export * from "./backgroundWatch";
 export * from "./constants";
 export * from "./ids";
+export * from "./display";
